Rename `install flux` to `enable repo`
