+++ pkg/ctl/enable/enable.go
@@ -11,6 +11,7 @@ func Command(flagGrouping *cmdutils.FlagGrouping) *cobra.Command {
 	verbCmd := cmdutils.NewVerbCmd("enable", "Enable features in a cluster", "")
 
 	cmdutils.AddResourceCmd(flagGrouping, verbCmd, enableProfileCmd)
+	cmdutils.AddResourceCmd(flagGrouping, verbCmd, enableRepo)
 
 	return verbCmd
 }
